+++ CHANGELOG.rst
@@ -2,9 +2,10 @@
 Changelog for package grepros
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-0.4.2 (2022-02-04)
+0.4.2 (2022-02-05)
 -------------------
 * add overwrite=true|false to --write options
+* add column-name=rostype:value to --write Parquet options
 * create message type definition from .idl if .msg file not available in ROS2
 
 0.4.1 (2022-01-08)